refactor(react-query): 로직 및 상수 분리
