+++ react-query/src/constants/api.ts
@@ -1,0 +1,5 @@
+export const API_BASE_URL = 'http://wooteco.com';
+
+export const API_ENDPOINTS = {
+  PRODUCTS: `${API_BASE_URL}/products`,
+};
