+++ react-query/src/constants/queryKeys.ts
@@ -1,0 +1,3 @@
+export const QUERY_KEYS = {
+  PRODUCTS: 'products',
+};
