+++ react-query/src/hooks/useProducts.ts
@@ -1,21 +1,10 @@
 import { useQuery } from '@tanstack/react-query';
-
-export interface Product {
-  id: number;
-  name: string;
-  price: number;
-  category: string;
-}
-
-async function fetchProducts(): Promise<Product[]> {
-  const response = await fetch('http://wooteco.com/products');
-
-  return await response.json();
-}
+import { fetchProducts } from '../api/products';
+import { QUERY_KEYS } from '../constants/queryKeys';
 
 export function useProducts() {
   return useQuery({
-    queryKey: ['products'],
+    queryKey: [QUERY_KEYS.PRODUCTS],
     queryFn: fetchProducts,
   });
 }
