+++ react-query/src/mocks/handlers.ts
@@ -1,8 +1,9 @@
 import { HttpResponse, http } from 'msw';
 import products from './products.json';
+import { API_ENDPOINTS } from '../constants/api';
 
 export const handlers = [
-  http.get('http://wooteco.com/products', () => {
+  http.get(API_ENDPOINTS.PRODUCTS, () => {
     return HttpResponse.json(products);
   }),
 ];
