+++ react-query/src/types/index.ts
@@ -1,0 +1,6 @@
+export interface Product {
+  id: number;
+  name: string;
+  price: number;
+  category: string;
+}
